fix another warning: scaled is not int
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2183 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/utils/utils.c
@@ -1477,7 +1477,7 @@ void checkpdfrestore(scaledpos pos)
     diff.h = pos.h - pos_stack[pos_stack_used].pos.h;
     diff.v = pos.v - pos_stack[pos_stack_used].pos.v;
     if (diff.h != 0 || diff.v != 0) {
-        pdftex_warn("Misplaced \\pdfrestore by (%dsp, %dsp)", diff.h, diff.v);
+        pdftex_warn("Misplaced \\pdfrestore by (%dsp, %dsp)", (int)diff.h, (int)diff.v);
     }
     if (page_mode) {
         matrix_stack_used = pos_stack[pos_stack_used].matrix_stack;
